requirements.txt
aioredis==1.2.0
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
asyncio==3.4.3
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.68.1
fastapi-limiter==0.1.4
greenlet==1.1.2
h11==0.12.0
hiredis==2.0.0
kombu==5.1.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dotenv==0.19.0
pytz==2021.3
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
